More refactoring
Moved all logic out of squareInput into crosswordContext.

Moved reducer function and associated types to the crossword reducer.
