+++ src/context/crosswordContext.tsx
@@ -1,14 +1,14 @@
 import { ReactNode, createContext, useContext, useReducer } from "react";
+import { KeyboardEvent, MouseEvent } from 'react';
 import Crossword from "../types/crossword";
 import Orientation from "../types/orientation";
-import UpdateGridCommand from "./updateGridCommand";
-import MovePositionCommand from "./movePositionCommand";
-import ToggleOrientationCommand from "./toggleOrientationCommand";
-import DeleteFillCommand from "./deleteFillCommand";
-import ToggleBlockCommand from "./toggleBlockCommand";
+import crosswordReducer, { CrosswordActions} from "./crosswordReducer";
 
 const defaultHeight = 15;
 const defaultWidth = 15;
+
+type SquareKeyDownEvent = KeyboardEvent<HTMLInputElement>;
+type SquareMouseEvent = MouseEvent<HTMLInputElement>;
 
 export enum MoveDirection {
   UP = 0,
@@ -20,11 +20,8 @@ export enum MoveDirection {
 
 export type CrosswordContextType = {
   crossword: Crossword,
-  updateFill: (x:number, y:number, value: string) => void,
-  movePosition: (x: number, y: number, direction: MoveDirection) => void,
-  toggleOrientation: () => void,
-  deleteFill: (x: number, y: number) => void,
-  toggleBlock: (x: number, y: number) => void,
+  onKeyDown: (e: SquareKeyDownEvent) => void
+  onClick: (e: SquareMouseEvent) => void
 };
 
 const CrosswordContext = createContext<CrosswordContextType|undefined>(undefined);
@@ -51,96 +48,9 @@ const initCrossword = (): Crossword => {
   };
 };
 
-export const CrosswordActions = {
-  updateFill: 'updateFill',
-  deleteFill: 'deleteFill',
-  movePosition: 'movePosition',
-  toggleOrientation: 'toggleOrientation',
-  toggleBlock: 'toggleBlock'
-};
-
-export type CrosswordActionPayload = {
-  x: number,
-  y: number,
-  value: string,
-};
-
-const crosswordReducer = (crossword: Crossword, action: {type: string, payload: CrosswordActionPayload}) => {
-  const payload = action.payload;
-  const actionType = action.type;
-
-  const getNextSquare = (x: number, y: number, orientation: Orientation):
-        [number, number] => {
-          return orientation === Orientation.across ?
-            [x, y + 1] : [x + 1, y];
-        };
-  const getPrevSquare = (x: number, y: number, orientation: Orientation):
-        [number, number] => {
-          return orientation === Orientation.across ?
-            [x, y - 1] : [x - 1, y];
-        };
-  const currentPos: [number, number] = [crossword.position.x, crossword.position.y];
-  const currentVal = crossword.grid[crossword.position.x][crossword.position.y];
-
-  switch(actionType) {
-  case CrosswordActions.updateFill: {
-    const nextSquare = getNextSquare(...currentPos, crossword.orientation);
-    const moveCmd = MovePositionCommand(...nextSquare, ...currentPos);
-
-    const updateCmd = UpdateGridCommand(payload.x, payload.y, payload.value,
-                                        currentVal);
-
-    // check if we are updating a block square
-    if(currentVal === '.') {
-      const toggleCmd = ToggleBlockCommand(payload.x, payload.y);
-      return moveCmd.do(updateCmd.do(toggleCmd.do(crossword)));
-    }
-
-    // otherwise update an move on
-    return moveCmd.do(updateCmd.do(crossword));
-  }
-  case CrosswordActions.movePosition: {
-    const moveCmd = MovePositionCommand(payload.x, payload.y, ...currentPos);
-    return moveCmd.do(crossword);
-  }
-  case CrosswordActions.toggleOrientation: {
-    const toggleCmd = ToggleOrientationCommand();
-    return toggleCmd.do(crossword);
-  }
-  case CrosswordActions.deleteFill: {
-    const prevSquare = getPrevSquare(...currentPos, crossword.orientation);
-    const moveCmd = MovePositionCommand(...prevSquare, ...currentPos);
-
-    // check if we are removing a block square
-    if(currentVal === '.') {
-      const toggleCmd = ToggleBlockCommand(payload.x, payload.y);
-      return moveCmd.do(toggleCmd.do(crossword));
-    }
-
-    // otherwise simply delete the square
-    const deleteCmd = DeleteFillCommand(payload.x, payload.y, currentVal);
-    return moveCmd.do(deleteCmd.do(crossword));
-  }
-  case CrosswordActions.toggleBlock: {
-    const nextSquare = getNextSquare(...currentPos, crossword.orientation);
-    const moveCmd = MovePositionCommand(...nextSquare,
-                                        crossword.position.x, crossword.position.y);
-    const toggleBlockCmd = ToggleBlockCommand(payload.x, payload.y);
-    return moveCmd.do(toggleBlockCmd.do(crossword));
-  }
-  default:
-      return crossword;
-  }
-};
-
 const CrosswordContextProvider = ({children}: {children: ReactNode}) => {
   const initState: Crossword = initCrossword();
   const [crosswordState, dispatch] = useReducer(crosswordReducer, initState);
-
-  const updateFill = (x: number, y: number, value: string) => {
-    dispatch({type: CrosswordActions.updateFill, payload:
-              { x: x, y: y, value: value }});
-  };
 
   const movePosition = (x: number, y: number, direction: MoveDirection) => {
     // if the movement direction is orthogonal to the current orientation
@@ -161,29 +71,51 @@ const CrosswordContextProvider = ({children}: {children: ReactNode}) => {
     }
   };
 
-  const toggleOrientation = () => {
-    dispatch({type: CrosswordActions.toggleOrientation,
+  const {x,y} = crosswordState.position;
+
+  const onKeyDown = (e: SquareKeyDownEvent) => {
+    const key = e.key.toUpperCase();
+
+    // all letters and numbers are valid
+    if(/[A-Z0-9]/.test(key) && e.key.length === 1) {
+      dispatch({type: CrosswordActions.updateFill, payload: {
+        x: x, y: y, value: key }});
+    } else if(key === '.') {
+      dispatch({type: CrosswordActions.toggleBlock,
+              payload: {x: x, y: y, value: ''}});
+    } else if(key === 'ARROWRIGHT') {
+      movePosition(x, y+1, MoveDirection.RIGHT);
+    } else if(key === 'ARROWLEFT') {
+      movePosition(x, y-1, MoveDirection.LEFT);
+    } else if(key === 'ARROWUP') {
+      movePosition(x-1, y, MoveDirection.UP);
+    } else if(key === 'ARROWDOWN') {
+      movePosition(x+1, y, MoveDirection.DOWN);
+    } else if(key === ' ') {
+      dispatch({type: CrosswordActions.toggleOrientation,
               payload: {x: 0, y: 0, value: ''}});
+    } else if (key === 'DELETE' || key === 'BACKSPACE') {
+      dispatch({type: CrosswordActions.deleteFill,
+              payload: {x: x, y: y, value: ''}});
+    }
   };
 
-  const deleteFill = (x: number, y: number) => {
-    dispatch({type: CrosswordActions.deleteFill,
-              payload: {x: x, y: y, value: ''}});
-  };
-
-  const toggleBlock = (x: number, y: number) => {
-    dispatch({type: CrosswordActions.toggleBlock,
-              payload: {x: x, y: y, value: ''}});
+  const onClick = (e: SquareMouseEvent) => {
+    const x: number = Number(e.currentTarget.attributes['data-xpos'].value);
+    const y: number = Number(e.currentTarget.attributes['data-ypos'].value);
+    if(e.type == 'click'){
+        movePosition(x,y,MoveDirection.JUMP);
+    } else if(e.type === 'dblclick') {
+      dispatch({type: CrosswordActions.toggleOrientation,
+                payload: {x: 0, y: 0, value: ''}});
+    }
   };
 
   return (
     <CrosswordContext.Provider value={{
       crossword: crosswordState,
-      updateFill: updateFill,
-      movePosition: movePosition,
-      toggleOrientation: toggleOrientation,
-      deleteFill: deleteFill,
-      toggleBlock: toggleBlock,
+      onKeyDown: onKeyDown,
+      onClick: onClick,
     }}>
         {children}
     </CrosswordContext.Provider>
